refactoring following PR comments
