+++ membership-attribute-service/conf/routes
@@ -9,7 +9,7 @@ GET         /user-attributes/me/mma-membership                          controll
 GET         /user-attributes/me/mma-monthlycontribution                 controllers.AccountController.monthlyContributionDetails
 GET         /user-attributes/me/mma-paper                               controllers.AccountController.paperDetails
 # this will become the dominant endpoint for getting subscription details
-GET         /user-attributes/me/mma                                     controllers.AccountController.allDetails
+GET         /user-attributes/me/mma                                     controllers.AccountController.allPaymentDetails
 
 POST        /user-attributes/me/membership-update-card                  controllers.AccountController.membershipUpdateCard
 POST        /user-attributes/me/digitalpack-update-card                 controllers.AccountController.digitalPackUpdateCard
